requirements.txt
PyYAML==5.4.1
torch==1.9.0
torchvision==0.10.0
matplotlib==3.4.2
